requirements.txt
langchain==0.3.2
langchain_community==0.3.1
langchain_core==0.3.9
langchain_groq==0.2.0
langchain_mistralai==0.2.0
langchain_openai==0.2.2
langchain_qdrant==0.1.4
mlflow==2.16.2
mlflow_skinny==2.16.2
pandas==2.2.3
python-dotenv==1.0.1
PyYAML==6.0.2
qdrant_client==1.11.3
ragas==0.1.21
